requirements.txt
streamlit>=1.30
sentence-transformers>=2.2.2
transformers>=4.36
torch>=2.0
faiss-cpu>=1.7.4
numpy>=1.24